requirements.txt
BTrees==6.1
cffi==1.17.1
persistent==6.1.1
pycparser==2.22
pycryptodome==3.23.0
pyscard==2.2.2
transaction==5.0
zc.lockfile==3.0.post1
ZConfig==4.2
ZODB==6.0.1
zodbpickle==4.2
zope.deferredimport==5.0
zope.interface==7.2
zope.proxy==6.1
